Add input check for use_data_object
The `use_data_object` requires a character vector as the input. However, previously no check was present to check this. This could result in incorrect additions to the datapackager.yml file. If the user accidentally provided the name of the object as a symbol rather than a character, then the elements of the object would be written to the yml file rather than the object itself.

This commit adds a check for a the input of `use_data_object` to be a character vector. In addition, since the package description for the function requires that the input have length 1, a check is made for this (but this restriction may be unnecessary).

Furthermore, tests that the input checks create errors when they should are added.

Changes to be committed:
	modified:   ../../R/use.R
	modified:   test-use_raw_data.R