Ensure that functions in R folder are sourced before building data-raw Rmd files
Previously the functions in R folder were not sourced before compiling data-raw Rmd files using `package_build`.
This meant that if any data-raw Rmd files depended on these functions, they would either throw an error
or the package would have to be installed before running package_build, which is inconvenient.

The functions in the R folder are now sourced before compiling data-raw Rmd files using `package_build`. Tests were written to ensure that this functionality is both required and that when it is used it runs successfully.

 On branch miguel_working
 Changes to be committed:
	modified:   ../../R/processData.R
	new file:   test-source_r_folder_functions.R